+++ docs/changes.rst
@@ -7,6 +7,8 @@ Serializable (don't include in main)
 * Defines the ``BaseStore`` interface for request storage mechanisms.
 * Added the setting ``TOOLBAR_STORE_CLASS`` to configure the request
   storage mechanism. Defaults to ``debug_toolbar.store.MemoryStore``.
+* Added setting ``SERIALIZATION_WARNINGS`` to log a warning when a
+  ``TypeError`` occurs during a panel's serialization.
 
 
 Pending